requirements.txt
torch==2.10.0
transformers==4.57.6
sentence-transformers==5.2.3
scikit-learn==1.8.0
joblib==1.5.3
numpy==2.4.2
pandas==3.0.1
fastapi==0.128.4
uvicorn==0.40.0
pydantic==2.12.5
chromadb==1.5.5
redis==7.3.0
google-genai>=0.8.0
python-dotenv==1.2.1
ollama>=0.4.0
sentencepiece==0.2.1